moved models into modules with names referencing the kategori of models.
