+++ src/folketinget/models/__init__.py
@@ -1,0 +1,7 @@
+from .sag import *
+from .sagstrin import *
+from .øvrige import *
+from .aktør import *
+from .dokument import *
+from .møde import *
+from .relationer import *
